fix create_token_array
